refactor!: use library pattern
