+++ lib/src/home_widget_info.dart
